+++ config/.babelrc.js
@@ -1,16 +1,17 @@
 module.exports = {
   presets: [
     [
-      require.resolve('babel-preset-env'),
+      require.resolve('@babel/preset-env'),
       {
         targets: {
-          node: '10',
+          node: '12',
         },
       },
     ],
   ],
   plugins: [
     require.resolve('babel-plugin-syntax-trailing-function-commas'),
-    require.resolve('babel-plugin-transform-object-rest-spread'),
+    require.resolve('@babel/plugin-proposal-object-rest-spread'),
+    require.resolve('babel-plugin-import-graphql'),
   ],
 };