+++ CHANGELOG.md
@@ -1,6 +1,44 @@
-# Change Log
+# Changelog
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
+
+## [2.0.0-beta.6](https://github.com/wasc-io/tools/compare/v2.0.0-beta.5...v2.0.0-beta.6) (2019-09-24)
+
+## [2.0.0-beta.5](https://github.com/wasc-io/tools/compare/v2.0.0-beta.4...v2.0.0-beta.5) (2019-09-24)
+
+
+### Features
+
+* **watch:** implement watch script via gulp-nodemon and gulp babel ([e168a34](https://github.com/wasc-io/tools/commit/e168a34))
+
+## [2.0.0-beta.4](https://github.com/wasc-io/tools/compare/v2.0.0-beta.3...v2.0.0-beta.4) (2019-09-24)
+
+
+### Bug Fixes
+
+* **format:** use correct stream for prettier formatting ([5d5aba0](https://github.com/wasc-io/tools/commit/5d5aba0))
+
+## [2.0.0-beta.3](https://github.com/wasc-io/tools/compare/v2.0.0-beta.2...v2.0.0-beta.3) (2019-09-23)
+
+
+### Features
+
+* switch prettier to stream based js-implementation ([7528197](https://github.com/wasc-io/tools/commit/7528197))
+
+## [2.0.0-beta.2](https://github.com/wasc-io/tools/compare/v2.0.0-beta.1...v2.0.0-beta.2) (2019-09-22)
+
+## [2.0.0-beta.1](https://github.com/wasc-io/tools/compare/v2.0.0-beta.0...v2.0.0-beta.1) (2019-09-22)
+
+## [2.0.0-beta.0](https://github.com/wasc-io/tools/compare/v1.25.1...v2.0.0-beta.0) (2019-09-22)
+
+
+### ⚠ BREAKING CHANGES
+
+* deprecation of build and watch subcommands
+
+### Features
+
+* simpler setup, no webpack, deprecation of subcommands, switch to yarn ([260f114](https://github.com/wasc-io/tools/commit/260f114))
 
 <a name="1.25.1"></a>
 ## [1.25.1](https://github.com/wasc-io/tools/compare/v1.25.0...v1.25.1) (2019-09-17)